+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.11.3] - 2023-06-26
+
+#### Added
+
+- resize and scroll events for the navigation bar
+
+#### Changed
+
+- a lot of work on the navbar and it's expansion when screen width is below `530px`
+  - removing all sidebar naming and labeling
+- messing around with some more css
+
 ## [0.11.2] - 2023-06-26
 
 #### Changed